refact: 테스트 위치 옮김
